cleanup

+++ Geo/MVertexBoundaryLayerData.cpp
@@ -7,28 +7,30 @@
 
 std::vector<MVertex*>* MVertexBoundaryLayerData::getChildren(int i)
 {
-  if (i < this->children.size() && i >= 0) {
+  if (i < (int)this->children.size() && i >= 0) {
     return &(children[i]);
-  } else {
+  }
+  else {
     return 0;
   }
 }
 
-int MVertexBoundaryLayerData::getNumChildren(int i) 
+int MVertexBoundaryLayerData::getNumChildren(int i)
 {
-  if (i < this->children.size() && i >= 0) {
+  if (i < (int)this->children.size() && i >= 0) {
     return this->children[i].size();
-  } else {
+  }
+  else {
     return -1;
   }
 }
 
-int MVertexBoundaryLayerData::getNumChildrenFamilies() 
+int MVertexBoundaryLayerData::getNumChildrenFamilies()
 {
   return this->children.size();
 }
 
-void MVertexBoundaryLayerData::addChildrenFamily(std::vector<MVertex*> family) 
+void MVertexBoundaryLayerData::addChildrenFamily(std::vector<MVertex*> family)
 {
   this->children.push_back(family);
 }
